Rearrange and cleanup
