+++ src/grammar/util/array_functions.h
@@ -1,0 +1,8 @@
+
+#ifndef ARRAY_FN_HEADER
+#define ARRAY_FN_HEADER
+
+int sum(int* arr, unsigned int len);
+int collapse(int* arr, unsigned int len);
+
+#endif
